Duplicate uprobe event detection.
We are seeing that on thread migration, uprobe events can sometime be duplicated.

The idea of the workaround is that for a given thread's sequence of u(ret)probe
events, two consecutive uprobe events must be associated with decreasing stack
pointers (nested function calls, stack grows by decreasing stack pointer).  If
we miss a uretprobe event, or if an extra uprobe event is generated, then the
second uprobe event will be associated with a stack pointer that is greater or
equal to the previous stack pointer, and that's not normal. In that situation,
we discard the second uprobe event.

My guess is that if a context switch happens exactly at the right time when the
uprobe's trampoline is being executed, the uprobe event will be generated on the
original core AND on the core we are migrating to.  This needs further
investigation.